university student management website project
In this project, I tried to design a site by focusing on the operation of functions using php at the beginner level. I used php html. The lecturer can actively use the functions of adding, deleting, listing, updating students on this site. In addition, features such as adding, updating, deleting and listing grades and courses for each student can be used actively and problem-free.
You can also see sample courses and students on the website.